ansible/requirements.txt
#request==0.0.0
request
ansible==2.2.0.0
docker-py==1.10.6
